Made several changes to the package
Currently we are able to get queries from CUIs, send those queries,
as well as create a dictionary of strings based on our CUIs. This is
all meant so that we can make good and concise graphs that show
relations between some concepts.

In the future, the point is to create a cohesive graph.